updated dtype in collate_fn

+++ transforms/transforms.py
@@ -52,8 +52,8 @@ def bds_collate_fn(batch):
     points, labels = zip(*batch)
     points = torch.stack(points)
     labels = torch.stack(labels)
-    point_batches = torch.zeros((points.shape[0] * points.shape[1], points.shape[2], points.shape[3]))
-    label_batches = torch.zeros((labels.shape[0] * labels.shape[1], labels.shape[2],))
+    point_batches = torch.zeros((points.shape[0] * points.shape[1], points.shape[2], points.shape[3]), dtype=torch.float32)
+    label_batches = torch.zeros((labels.shape[0] * labels.shape[1], labels.shape[2],), dtype=torch.long)
     for i in range(points.shape[0]):
         for j in range(points.shape[1]):
             point_batches[i * j] = points[i][j]
